% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hackspatialcpp-package.R
\docType{package}
\name{hackspatialcpp-package}
\alias{hackspatialcpp}
\alias{hackspatialcpp-package}
\title{hackspatialcpp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/diminutive/hackspatialcpp}
  \item Report bugs at \url{https://github.com/diminutive/hackspatialcpp/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
